Fixed regex (original one didn't work for the examples given) and added option to use a dictionary-based SIT
The regex given in the example didn't match the example values given, so it would not detect them as stated. 
Also, using a dictionary instead of a regex is becoming a more popular alternative than a regex for multi-word evidence scenarios, so updated the explanation to center on that alternative.